Public API

Objects

TclTk.TclObj — Type

TclObj(val) -> obj

Return a Tcl object storing value val. The initial type of the Tcl object, given by obj.type, depends on the type of val:

A string, symbol, or character is stored as a Tcl :string.

A Boolean or integer is stored as a Tcl :int or :wideInt.

A non-integer real is stored as a Tcl :double.

A dense vector of bytes (UInt8) is stored as a Tcl :bytearray.

A tuple is stored as a Tcl :list.

A Tcl object is returned unchanged. Call copy to have an independent copy.

If the content of a Tcl object is valid as a list, the object may be indexed, elements may be added, deleted, etc.

Properties

Tcl objects have the following read-only properties:

obj.refcnt is the reference count of obj. If obj.refcnt > 1, the object is shared and must be copied before being modified.

obj.ptr is the pointer to the Tcl object, this is the same as pointer(obj).

obj.type is the current internal type of obj as a symbol. This type may change depending on how the object is used by Tcl. For example, after having evaluated a script in a Tcl string object, the object internal state becomes :bytecode to reflect that it now stores compiled byte code.

Conversion

Call convert(T, obj) to get a value of type T from Tcl object obj. The content of a Tcl object may always be converted into a string by calling convert(String, obj), string(obj), or String(obj) which all yield a copy of this string.

See also

tcl_list or tcl_concat for building Tcl objects to efficiently store arguments of Tcl commands.

Methods TclTk.Impl.new_object and TclTk.Impl.unsafe_convert may be extended to convert other types of value to Tcl object.

TclTk.tcl_concat — Function

tcl_concat(args...) -> list::TclObj

Return a list of Tcl objects obtained by concatenating the elements of the arguments arg... each being considered as a list. This mimics the behavior of the Tcl concat command.

See also

tcl_list, tcl_eval, and TclObj.

TclTk.tcl_list — Function

 tcl_list(args...) -> list::TclObj

Return a list of Tcl objects such that each of args... is a single element of the returned list. This mimics the behavior of the Tcl list command.

See also

tcl_concat, tcl_eval, and TclObj.

Interpreters

TclTk.tcl_eval — Function

tcl_eval(T=Nothing, args...) -> res::T

Concatenate arguments args... into a list, evaluate this list as a Tcl script, and return a value of type T.

Except for the specific handling of args..., the returned result follows the same behavior as tcl_exec.

Examples

With tcl_eval, storing a value in a global variable can be done by:

julia> tcl_eval(TclObj, "set x 42")
TclObj("42")

whose result is a Tcl object storing the string "42".

See also

See tcl_concat for the rules to concatenate arguments into a list (apart from the accounting of pairs).

See tcl_exec for another way to execute a Tcl command where each of args... is considered as a distinct command argument.

tcl_eval(T=Nothing, script::AbstractString) -> res::T

Evaluate a Tcl script provided as a string. It returns the same result as tcl_eval.

TclTk.tcl_exec — Function

tcl_exec(T=Nothing, args...) -> res::T

Make a list out of the arguments args..., evaluate this list as a Tcl command, and return a value of type T. Any key => val pair in args... and keywords in kwds... is converted in the pair of arguments -key and val in the command list (where the leading hyphen before the key name is added if needed).

The evaluation of a Tcl command stores a result (or an error message) in the Tcl interpreter and returns a status. The behavior of tcl_exec depends on the type T of the expected result:

If T is Tuple{TclStatus,R}, the status and the result of the Tcl command are returned as a 2-tuple and with the result converted to type R. In practice, R is one of TclObj, String, or Nothing. No conversion of the result is attempted if R is Nothing which is useful when the caller is only interested in the status.

Otherwise, if the command status is TCL_OK, the result of the command is returned as a value of type T. No conversion of the result is attempted if T is Nothing.

Otherwise, if the command status is not TCL_OK, a TclError exception is thrown.

Examples

With tcl_exec, storing a value in a global variable can be done by:

julia> tcl_exec(TclObj, :set, :x, 42)
TclObj(42)

whose result is a Tcl object storing the integer 42.

See also

See tcl_list for the rules to build a list (apart from the accounting of pairs).

See tcl_eval for another way to evaluate a Tcl script. The difference with tcl_eval is that each input argument is interpreted as a different token of the Tcl command and the accounting of key => val pairs.

TclTk.tcl_quote_string — Function

tcl_quote_string(str)

Return string str a valid Tcl string surrounded by double quotes that can be directly inserted in Tcl scripts.

This is similar to escape_string but specialized to represent a valid Tcl string surrounded by double quotes in a script.

Status and exceptions

TclTk.TclStatus — Type

TclStatus

Type of result returned by evaluating Tcl scripts or commands. Possible values are:

TCL_OK: Command completed normally; the interpreter's result contains the command's result.

TCL_ERROR: The command couldn't be completed successfully; the interpreter's result describes what went wrong.

Other existing values only used internally:

TCL_RETURN: The command requests that the current function return; the interpreter's result contains the function's return value.

TCL_BREAK: The command requests that the innermost loop be exited; the interpreter's result is meaningless.

TCL_CONTINUE: Go on to the next iteration of the current loop; the interpreter's result is meaningless.

TclTk.TclError — Type

TclError(args...)

Return a TclError exception with error message given by string(args...).

TclTk.tcl_error — Function

tcl_error(args...)

Throw a TclError exception with error message given by string(args...).

Variables

TclTk.TclArray — Type

A = TclArray{K,V}(arr)

Return an abstract dictionary A which is linked to a global Tcl array named arr. Expression A[key] refer to the element key in the global Tcl array arr which takes the form arr(key) in a Tcl script.

Key and value types, K and V, default to TclObj if not specified.

Property A.name yields the name of the global Tcl array.

Examples

julia> A = TclArray{String,Real}("data", "index" => 7, "value" => 33.5)
TclArray{String,Real}("::data") with 2 entries:
  "index" => 7
  "value" => 33.5

julia> A["index"]
7

See also

tcl_isassigned, tcl_getvar, tcl_setvar, tcl_unsetvar, and TclVariable.

TclTk.TclVariable — Type

A = TclVariable{T}(var)

Return a Julia object A linked to the global Tcl variable named var. The variable value is given by A[] and can be set by A[] = x.

Type parameter T is the assumed Julia type of the value that can be stored in this variable. If not specified, TclObj is assumed for T, hence the variable may store any type of Tcl object. If variable type parameter is T = TclObj, expression fetch(S, A), with S a type, can be used to efficiently convert the value of the variable to type S.

Currently, var must refer to a simple Tcl variable or to a single element of a Tcl array (in which case, name may be a 2-tuple (part1, part2)), not to a Tcl array name.

Property A.name yields the name of the global Tcl variable.

Call eltype(A) to retrieve the type of A and isassigned(A) or tcl_isassigned(A) to check whether A has an associated value. To unset the value of A, call delete!(A), tcl_unsetvar(A), or do A[] = unset.

Examples

julia> A = TclVariable{Int}("::GLOBAL_COUNTER")
TclVariable{Int64}(name: "::GLOBAL_COUNTER", value: #undef)

julia> A[] = 0
0

julia> A
TclVariable{Int64}(name: "::GLOBAL_COUNTER", value: 0)

julia> tcl_exec(:incr, A.name, 4) # increment variable with Tcl `incr` command

julia> A[]
4

See also

tcl_isassigned, tcl_getvar, tcl_setvar, tcl_unsetvar, and TclArray.

TclTk.tcl_getvar — Function

tcl_getvar(T=TclObj, var) -> val::T

Return the value of the global Tcl variable named var. For a Tcl array element, var may be a 2-tuple of array and element names.

Optional argument T (TclObj by default) can be used to specify the type of the returned value. Some possibilities are:

If T is TclObj (the default), a managed Tcl object is returned. This is the most efficient if the returned value is intended to be used in a Tcl list or as an argument of a Tcl script or command.

If T is Bool, a boolean value is returned.

If T is String, a string is returned.

If T is Char, a single character is returned (an exception is thrown if Tcl object is not a single character string).

If T <: Integer, an integer value of type T is returned.

If T <: AbstractFloat, a floating-point value of type T is returned.

Note that, except if T is TclObj, a conversion of the Tcl object stored by the variable may be needed.

If T is unspecified, var may be an instance of TclVariable.

See also

tcl_isassigned, tcl_setvar, tcl_unsetvar, and TclVariable.

TclTk.tcl_isassigned — Function

tcl_isassigned(var) -> bool::Bool

Return whether global Tcl variable named var is associated with a value. For a Tcl array element, var may be a 2-tuple (part1, part2).

See also

tcl_getvar, tcl_setvar, tcl_unsetvar, and TclVariable.

TclTk.tcl_setvar — Function

tcl_setvar(var, val) -> nothing
tcl_setvar(T, var, val) -> newvalue::T

Set global Tcl variable var to value val. For a Tcl array element, var may be a 2-tuple of array and element names.

The Tcl variable is deleted if val is unset, the singleton provided by the UnsetIndex package and exported by the Tcl package.

If a leading argument T other than Nothing is specified, the new value of the variable is returned as an instance of type T (can be TclObj). The new value may be different from val because of trace(s) associated to this variable.

See also

tcl_getvar, tcl_isassigned, tcl_unsetvar, and TclVariable.

TclTk.tcl_unsetvar — Function

tcl_unsetvar(var)

Delete global Tcl variable named var. For a Tcl array element, var may be a 2-tuple of array and element names.

Keywords

Keyword nocomplain can be set true to ignore errors. By default, nocomplain=false.

See also

tcl_getvar, tcl_isassigned, tcl_setvar, and TclVariable.

Callbacks

TclCallback
TclTk.deletecommand

Widgets

TclTk.TkWidget — Type

TkWidget(path)

Return a widget for the given Tk window path. The type of the widget is inferred from the class of the Tk window.

TclTk.Impl.Ttk.Button — Type

w = Button(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a button widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing button widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing button widget in a Button structure:

w = Button(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Tk.Canvas — Type

w = Canvas(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a canvas widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing canvas widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing canvas widget in a Canvas structure:

w = Canvas(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Checkbutton — Type

w = Checkbutton(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a check button widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing check button widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing check button widget in a Checkbutton structure:

w = Checkbutton(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Combobox — Type

w = Combobox(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a combo box widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing combo box widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing combo box widget in a Combobox structure:

w = Combobox(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Entry — Type

w = Entry(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap an entry widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing entry widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing entry widget in a Entry structure:

w = Entry(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Frame — Type

w = Frame(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a frame widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing frame widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing frame widget in a Frame structure:

w = Frame(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Label — Type

w = Label(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a label widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing label widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing label widget in a Label structure:

w = Label(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Labelframe — Type

w = Labelframe(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a label frame widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing label frame widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing label frame widget in a Labelframe structure:

w = Labelframe(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Tk.Listbox — Type

w = Listbox(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a list box widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing list box widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing list box widget in a Listbox structure:

w = Listbox(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Tk.Menu — Type

w = Menu(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a menu widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing menu widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing menu widget in a Menu structure:

w = Menu(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Menubutton — Type

w = Menubutton(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a menu button widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing menu button widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing menu button widget in a Menubutton structure:

w = Menubutton(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Tk.Message — Type

w = Message(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a message widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing message widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing message widget in a Message structure:

w = Message(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Notebook — Type

w = Notebook(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a notebook widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing notebook widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing notebook widget in a Notebook structure:

w = Notebook(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Panedwindow — Type

w = Panedwindow(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a paned window widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing paned window widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing paned window widget in a Panedwindow structure:

w = Panedwindow(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Progressbar — Type

w = Progressbar(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a progress bar widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing progress bar widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing progress bar widget in a Progressbar structure:

w = Progressbar(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Radiobutton — Type

w = Radiobutton(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a radio button widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing radio button widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing radio button widget in a Radiobutton structure:

w = Radiobutton(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Scale — Type

w = Scale(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a scale widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing scale widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing scale widget in a Scale structure:

w = Scale(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Scrollbar — Type

w = Scrollbar(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a scroll bar widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing scroll bar widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing scroll bar widget in a Scrollbar structure:

w = Scrollbar(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Separator — Type

w = Separator(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a separator widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing separator widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing separator widget in a Separator structure:

w = Separator(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Sizegrip — Type

w = Sizegrip(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a size grip widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing size grip widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing size grip widget in a Sizegrip structure:

w = Sizegrip(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Ttk.Spinbox — Type

w = Spinbox(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a spin box widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing spin box widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing spin box widget in a Spinbox structure:

w = Spinbox(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Tk.Text — Type

w = Text(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a text widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing text widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing text widget in a Text structure:

w = Text(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.Impl.Tk.Toplevel — Type

w = Toplevel([path,] pairs::Pair...; kwds...)

Create or wrap a top-level Tk widget.

Optional argument path is the full path of the widget relative to its parent path; if not specified, a unique path is automatically generated. If path is specified and corresponds to an existing top-level widget, no new widget is created and the existing widget is wrapped in Toplevel structure.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. Function TclTk.configure may be called to retrieve a list of possible options.

For example:

top = Toplevel()

creates a new top-level widget in while

main = Toplevel(".")

returns the main Tk window.

TclTk.Impl.Ttk.Treeview — Type

w = Treeview(parent::Widget, [name,] pairs::Pair...; kwds...)

Create or wrap a tree view widget in parent widget.

Optional argument name is the name the widget relative to its parent path; if not specified, a unique name is automatically generated. If name is specified and corresponds to an existing tree view widget, no new widget is created and the existing widget is wrapped as explained below.

Any number of configurable options may be specified by opt => val pairs of by opt=val keywords with opt the option name and val the option value. The option name has no leading hyphen and, if the option is specified by a pair, is a string or a symbol. TclTk.configure may be called to retrieve a list of possible options.

The constructor may also wrap an existing tree view widget in a Treeview structure:

w = Treeview(path, pairs::Pair...; kwds...)

where path is the full path of the widget and pairs... denotes optional configurable options which are applied to the existing widget.

TclTk.cget — Function

TclTk.cget(T=TclObj, w, opt) -> val::T

Return the value, converted to type T, of the option opt for Tk object (widget or image) w. Option opt may be specified as a string or as a Symbol and shall corresponds to a Tk option name (the leading hyphen may be omitted). Another way to obtain an option value is:

w[opt] -> val::TclObj

See also

TclTk.configure and TkWidget.

TclTk.configure — Function

TclTk.configure(w)
w.configure()
w(:configure)

Return all the options of Tk object (widget or image) w.

TclTk.configure(w, pairs...; kwds...)
w.configure(pairs...; kwds...)
w(:configure, pairs...; kwds...)

Change some options of widget or image w. Trailing pairs... arguments and keywords kwds... are interpreted as configuration options. Another way to change the settings is:

w[opt1] = val1
w[opt2] = val2

See also

TclTk.cget and TkWidget.

TclTk.winfo — Function

TclTk.winfo(T=TclObj, option, args...) -> res::T
TclTk.winfo(T=TclObj, option, w::TkWidget) -> res::T
TclTk.winfo(T=TclObj, w::TkWidget, option) -> res::T

Return information related to Tk window(s) or to specific Tk widget w as a value of type T.

The returned information depends on option, one of atom, atomname, cells, children, class, colormapfull, containing, depth, exists, fpixels, geometry, height, id, interps, ismapped, manager, name, parent, path, pathname, pixels, pointerx, pointerxy, pointery, reqheight, reqwidth, rgb, rootx, rooty, screen, screencells, screendepth, screenheight, screenmmheight, screenmmwidth, screenvisual, screenwidth, server, toplevel, viewable, visual, visualid, visualsavailable, visualsavailable_includeids, vrootheight, vrootwidth, vrootx, vrooty, width, x, or y.

Window and geometry managers

TclTk.grid — Function

TclTk.grid(T=Nothing, args...)

Call Tk grid geometry manager. One of the arguments must be a widget (that is an instance of TkWidget). Optional argument T is to specify the result type.

To specify the grid manager options for a single widget w, another possible syntax is:

w.grid(T=Nothing, args...; kwds...)

See also

TclTk.pack and TclTk.place.

TclTk.pack — Function

TclTk.pack(T=Nothing, args...; kwds...)

Call Tk packer geometry manager. One of the arguments must be a widget (that is an instance of TkWidget). Optional argument T is to specify the result type.

To specify the packing options for a single widget w, another possible syntax is:

w.pack(T=Nothing, args...; kwds...; kwds..)

For example:

using TclTk
top = Toplevel()
wm.title(top, "A simple example")
btn = Button(top, text="Click me", command="puts {ouch!}")
btn.pack(side=:bottom, padx=30, pady=5)

See also

TclTk.grid and TclTk.place.

TclTk.place — Function

TclTk.place(T=Nothing, args...; kwds..)

Call Tk placer geometry manager. One of the arguments must be a widget (that is an instance of TkWidget). Optional argument T is to specify the result type.

To specify the placing options for a single widget w, another possible syntax is:

w.place(args...; kwds...)

See also

TclTk.grid and TclTk.pack.

TclTk.wm — Function

TclTk.wm(T=Nothing, cmd, w::TkWidget, args...; kwds...) -> res::T
wm.cmd(T=Nothing, w::TkWidget, args...; kwds...) -> res::T

Interact with the window manager to query or control such things as the title for widget w, its geometry, etc. Argument T is the expected type for the result. With the syntax wm.cmd(w, ...) the result has a suitable default type that depends on cmd.

The window manager command cmd is one of aspect, attributes, client, colormapwindows, command, deiconify, focusmodel, forget, frame, geometry, grid, group, iconbadge, iconbitmap, iconify, iconmask, iconname, iconphoto, iconposition, iconwindow, manage, maxsize, minsize, overrideredirect, positionfrom, protocol, resizable, sizefrom, stackorder, state, title, transient, or withdraw.

Images

TclTk.TkImage — Type

TkImage{type}(pairs...; kwds...) -> img
TkImage{type}(name, pairs...; kwds...) -> img

Return a Tk image of given type (e.g., :bitmap or :photo).

If the image name is omitted, it is automatically generated. If name is specified and an image with this name already exists, it is re-used and, if options are specified, it is reconfigured.

Trailing pairs... arguments and keywords kwds... are interpreted as image configuration options. Possible options depend on the image types.

A Tk image can then be used in any Tcl/Tk script or command where an image is expected.

Tk images implement the abstract array API. To extract the pixels of an image, the img[x,y] syntax may be used with x and y pixel indices or ranges. For a :photo image, pixel values are represented by RGBA{N0f8} colors.

A Tk image has a number of properties:

img.inuse    # whether an image is in use in a Tk widget
img.width    # the width of the image in pixels
size(img, 1) # idem
img.height   # the height of the image in pixels
size(img, 2) # idem
img.size     # (width, height)
size(img)    # idem
img.type     # the symbolic type of the image (`:bitmap`, `:photo`, etc.)
img.name     # the image name
img.handle   # the image handle

See also

TkBitmap and TkPhoto are aliases for specific image types.

TkImage(name, pairs...; kwds...) -> img

Return an instance of TkImage managing Tk image named name and after applying any options specified by pairs... and kwds....

TclTk.TkBitmap — Type

TkBitmap(args...) -> img
TkImage{:bitmap}(args...) -> img

Return a Tk bitmap image. See TkImage for more information.

TclTk.TkPhoto — Type

TkPhoto(args...) -> img
TkImage{:photo}(args...) -> img

Return a Tk photo image. See TkImage for more information.